pyproject.toml
[project]
name = "heuristos"
version = "0.1.0"
description = "Automatic Analysis of Legal Agreements"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "attrs>=25.1.0",
    "beautifulsoup4>=4.13.3",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "ordered-set>=4.1.0",
]

[dependency-groups]
dev = [
    "isort>=6.0.0",
]

[build-system]
requires = ["setuptools>=75.8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["heuristos"]
package-dir = {"" = "src"}
